backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-multipart
passlib[bcrypt]
pyjwt
python-dotenv
requests
pandas
tabulate
